prepare for production chunking
