+++ crates/next-core/src/next_edge/context.rs
@@ -6,7 +6,10 @@ use turbo_tasks_fs::FileSystemPath;
 use turbopack::resolve_options_context::ResolveOptionsContext;
 use turbopack_browser::BrowserChunkingContext;
 use turbopack_core::{
-    chunk::{module_id_strategies::ModuleIdStrategy, ChunkingContext, MinifyType},
+    chunk::{
+        module_id_strategies::ModuleIdStrategy, ChunkingContext, EcmascriptChunkingConfig,
+        MinifyType,
+    },
     compile_time_info::{
         CompileTimeDefineValue, CompileTimeDefines, CompileTimeInfo, DefineableNameSegment,
         FreeVarReference, FreeVarReferences,
@@ -211,29 +214,32 @@ pub async fn get_edge_chunking_context_with_client_assets(
 ) -> Result<Vc<Box<dyn ChunkingContext>>> {
     let output_root = node_root.join("server/edge".into()).to_resolved().await?;
     let next_mode = mode.await?;
-    Ok(Vc::upcast(
-        BrowserChunkingContext::builder(
-            root_path,
-            output_root,
-            output_root_to_root_path,
-            client_root,
-            output_root.join("chunks/ssr".into()).to_resolved().await?,
-            client_root
-                .join("static/media".into())
-                .to_resolved()
-                .await?,
-            environment,
-            next_mode.runtime_type(),
-        )
-        .asset_base_path(asset_prefix)
-        .minify_type(if *turbo_minify.await? {
-            MinifyType::Minify
-        } else {
-            MinifyType::NoMinify
-        })
-        .module_id_strategy(module_id_strategy)
-        .build(),
-    ))
+    let mut builder = BrowserChunkingContext::builder(
+        root_path,
+        output_root,
+        output_root_to_root_path,
+        client_root,
+        output_root.join("chunks/ssr".into()).to_resolved().await?,
+        client_root
+            .join("static/media".into())
+            .to_resolved()
+            .await?,
+        environment,
+        next_mode.runtime_type(),
+    )
+    .asset_base_path(asset_prefix)
+    .minify_type(if *turbo_minify.await? {
+        MinifyType::Minify
+    } else {
+        MinifyType::NoMinify
+    })
+    .module_id_strategy(module_id_strategy);
+
+    if !next_mode.is_development() {
+        builder = builder.ecmascript_chunking_config(EcmascriptChunkingConfig {})
+    }
+
+    Ok(Vc::upcast(builder.build()))
 }
 
 #[turbo_tasks::function]
@@ -248,28 +254,31 @@ pub async fn get_edge_chunking_context(
 ) -> Result<Vc<Box<dyn ChunkingContext>>> {
     let output_root = node_root.join("server/edge".into()).to_resolved().await?;
     let next_mode = mode.await?;
-    Ok(Vc::upcast(
-        BrowserChunkingContext::builder(
-            root_path,
-            output_root,
-            node_root_to_root_path,
-            output_root,
-            output_root.join("chunks".into()).to_resolved().await?,
-            output_root.join("assets".into()).to_resolved().await?,
-            environment,
-            next_mode.runtime_type(),
-        )
-        // Since one can't read files in edge directly, any asset need to be fetched
-        // instead. This special blob url is handled by the custom fetch
-        // implementation in the edge sandbox. It will respond with the
-        // asset from the output directory.
-        .asset_base_path(ResolvedVc::cell(Some("blob:server/edge/".into())))
-        .minify_type(if *turbo_minify.await? {
-            MinifyType::Minify
-        } else {
-            MinifyType::NoMinify
-        })
-        .module_id_strategy(module_id_strategy)
-        .build(),
-    ))
+    let mut builder = BrowserChunkingContext::builder(
+        root_path,
+        output_root,
+        node_root_to_root_path,
+        output_root,
+        output_root.join("chunks".into()).to_resolved().await?,
+        output_root.join("assets".into()).to_resolved().await?,
+        environment,
+        next_mode.runtime_type(),
+    )
+    // Since one can't read files in edge directly, any asset need to be fetched
+    // instead. This special blob url is handled by the custom fetch
+    // implementation in the edge sandbox. It will respond with the
+    // asset from the output directory.
+    .asset_base_path(ResolvedVc::cell(Some("blob:server/edge/".into())))
+    .minify_type(if *turbo_minify.await? {
+        MinifyType::Minify
+    } else {
+        MinifyType::NoMinify
+    })
+    .module_id_strategy(module_id_strategy);
+
+    if !next_mode.is_development() {
+        builder = builder.ecmascript_chunking_config(EcmascriptChunkingConfig {})
+    }
+
+    Ok(Vc::upcast(builder.build()))
 }
